+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.5+1
+
+* Adds remaining methods for internal wrapper of the Android native `AdsRequest`.
+* Adds remaining methods for internal wrapper of the iOS native `IMAAdsRequest`.
+
 ## 0.2.5
 
 * Adds support to set general SDK settings. See `ImaSettings` and `AdsLoader.settings`.